Move generic handler up out of linode package
